Some versions of the RevPi Gate firmware break if they receive malformed packets (too large, too short, wrong ethertype, non-broadcast): Such packets cause the firmware to not drain its receive queue. Over time, packets accumulate in the queue and the firmware always responds to an older packet instead of the one it has just received.
Since commit 2de8667 ("piControl: Use native network stack for
RevPi Gate communication"), the ks8851 chips on the RevPi Core and
Connect are exposed as regular Ethernet interfaces, which results in
occasional IP packets sent to attached gateways by dhcpcd and avahi,
reproducibly causing their firmware to break.

Although the bug has been fixed in the firmware, updating it on all
affected gateways will take time.  As a workaround, register a netfilter
hook upon receiving the first packet from a gateway such that egress
packets on that interface are dropped unless their ethertype corresponds
to the one used for the Gateway protocol, which is 0x419c.