piControl: Recognize RevPi Connect Flat
If the "compatible" string at the DeviceTree's apex does not specify a
known RevPi model, piControl assumes that it is running on a RevPi Core.
This enables backward compatibility for machines shipped in the wheezy
and jessie era which lacked a RevPi model in the "compatible" string.

As a result, the new RevPi Connect Flat is incorrectly deemed a Core,
with the undesirable side effect that GPIO 42 and 43 are manipulated by
piControl upon probing, which are used for spi2 on the Connect Flat.

Avoid by recognizing the Connect Flat as a first step towards properly
supporting it in piControl.

Signed-off-by: Lukas Wunner <lukas@wunner.de>